// src/model/etudiant.ts
export interface Etudiant {
    id: number;
    nom_etudiant: string;
    prenom_etudiant: string;
    email: string;
    mot_de_passe: string;
    filiereId: number;
    promotionId: number;
    faculteId: number;
    departementId: number;
    userId: number;
    createdAt: string;
    updatedAt: string;

}